+++ peerinst/templates/registration/sign_up_admin_email_html.html
@@ -18,7 +18,7 @@
 
 <p>{% trans "Access your administrator account to activate this new user at:" %}</p>
 
-<p>https://{{ request.get_host }}{% url 'admin--new-user-approval' %}</p>
+<p>https://{{ request.get_host }}{% url 'saltise-admin:new-user-approval' %}</p>
 
 <p>{% trans 'Cheers'%},</p>
